pyproject.toml
[tool.poetry]
name = "cishe"
version = "0.1.0"
description = "Client Information System for HelloEdu"
authors = ["Ace <ace.jl.han@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.*"
arrow = "^0.15.8"
djangorestframework = "^3.11.1"
requests = "^2.24.0"
openpyxl = "^3.0.5"
django-redis = "^4.12.1"
pyyaml = "^5.3.1"
django-memoize = "^2.3.1"
pymysql = {version = "^0.10.0", extras = ["rsa"]}
django-environ = "^0.4.5"
django-phonenumber-field = "^5.0.0"
djangorestframework-simplejwt = {git = "https://github.com/AtuzSolution/django-rest-framework-simplejwt", rev = "jwt_cookie"}
drf-flex-fields = "^0.8.5"


[tool.poetry.dev-dependencies]
# code formatter
black = "^19.10b0"
ptvsd = "^4.3.2"
pytest-django = "^3.9.0"
pytest-mock = "^3.2.0"
pytest-xdist = "^2.0.0"
tox = "^3.19.0"
pytest-cov = "^2.10.1"
pytest-html = "^2.1.1"
# pytest-factoryboy = "^2.0.3"

# code checker
flake8 = "^3.8.3"
flake8-docstrings = "^1.5.0"
isort = "^5.4.2"
seed-isort-config = "^2.2.0"
pre-commit = "^2.6.0"
pydocstyle = "^5.0.2"
django-stubs = "^1.5.0"
djangorestframework-stubs = "^1.2.0"
mypy = "^0.770"
codecov = "^2.1.9"

[tool.black]
target_version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git/
  |\.eggs
  |__pycache__
  |\.cache
  |\.ipynb_checkpoints
  |\.mypy_cache
  |\.pytest_cache
  |\.tox
  |dist
  |media
  |templates
)/
'''

[tool.isort]
known_first_party = ''
known_third_party = ["django", "environ", "phonenumber_field", "rest_flex_fields", "rest_framework", "rest_framework_simplejwt"]
multi_line_output = 3
lines_after_imports = 2
force_grid_wrap = 0
combine_as_imports = true
include_trailing_comma = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
